+++ spec/Method/Payment/CreateSpec.php
@@ -10,6 +10,26 @@ class CreateSpec extends ObjectBehavior
 
     function let()
     {
+        $browserInfo = [
+            'accept_header' => 'Some header',
+            'browser_language' => 'LT',
+            'screen_width' => 390,
+            'screen_height' => 400,
+            'challenge_window_size' => '390x400',
+            'user_agent' => 'super user agent',
+            'color_depth' => 24,
+            'time_zone' => -60,
+            // 'ip_address' => '192.168.0.1',
+        ];
+        $billingAddress = [
+            'address_line1' => 'first address line',
+            'city' => 'balbieriskis',
+            'country' => 'LT',
+            'postal_code' => '0234'
+        ];
+        $threeds2data['notification_url'] = 'http://localhost:8000/3dsv2_callback.php';
+        $threeds2data['browser_info'] = $browserInfo;
+
         $this->options = [
             'amount' => 12.99,
             'currency' => 'EUR',
@@ -25,6 +45,7 @@ class CreateSpec extends ObjectBehavior
                 'cvc' => '456',
                 'holder' => 'Mr Tester',
             ],
+            'threeds2_data' => $threeds2data
         ];
         $this->beConstructedWith($this->options);
     }